Rename file with main string function
