+++ rebdev.pavel/P3/main.cpp
@@ -3,7 +3,7 @@
 #include <cstddef>
 #include <cstring>
 #include "stringInput.hpp"
-#include "twoStrings.hpp"
+#include "stringConvert.hpp"
 
 
 int main()
